Fixed build error

+++ src/main/java/io/swagger/models/properties/PropertyBuilder.java
@@ -6,6 +6,7 @@ import io.swagger.models.Model;
 import io.swagger.models.ModelImpl;
 import io.swagger.models.RefModel;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -672,11 +673,11 @@ public class PropertyBuilder {
 
         protected <N extends AbstractNumericProperty> N mergeNumeric(N property, Map<PropertyId, Object> args) {
             if (args.containsKey(PropertyId.MINIMUM)) {
-                final Double value = PropertyId.MINIMUM.findValue(args);
+            	final BigDecimal value = PropertyId.MINIMUM.findValue(args);
                 property.setMinimum(value);
             }
             if (args.containsKey(PropertyId.MAXIMUM)) {
-                final Double value = PropertyId.MAXIMUM.findValue(args);
+            	final BigDecimal value = PropertyId.MAXIMUM.findValue(args);
                 property.setMaximum(value);
             }
             if (args.containsKey(PropertyId.EXCLUSIVE_MINIMUM)) {
